refactor(hpack): widen shared prefixed integers to u64
QPACK (RFC 9204 Section 4.1.1) requires decoding integers up to 62
bits long, while HPACK only needs 32; both use the same RFC 7541
Section 5.1 representation. Widen encode/decode to u64 so the shared
primitive serves both protocols; HPACK call sites cast to u64/u32
unchanged in behavior.

The 62-bit limit is enforced on the decoded value, not on the octet
shift: the ninth continuation octet (shift 56) already carries bit
62, so a shift-based guard would admit 63-bit values. Values >= 2^62
are rejected with HpackError::InvalidInteger.

Adds round-trip coverage for the 62-bit boundary, rejection of 63-
and 64-bit values, and 2..=8-bit prefix sizes (prefixes HPACK never
uses).

Verification: cargo test --features h2 --lib hpack (61 passing),
cargo test --features h3 --lib hpack (61 passing),
cargo clippy --features h3 -- -D warnings, cargo fmt --check.

+++ src/h2/hpack/encode.rs
@@ -86,14 +86,14 @@ impl Encoder {
     pub fn encode(&mut self, headers: &[Header], out: &mut Vec<u8>) {
         if let Some(size) = self.queued_size_update.take() {
             self.table.set_max_size(size);
-            integer::encode(out, size as u32, 5, SIZE_UPDATE);
+            integer::encode(out, size as u64, 5, SIZE_UPDATE);
         }
 
         for header in headers {
             let name = header.name();
             if NEVER_INDEXED.contains(&name) {
                 match self.table.find_name(name) {
-                    Some(index) => integer::encode(out, index as u32, 4, LITERAL_NEVER_INDEXED),
+                    Some(index) => integer::encode(out, index as u64, 4, LITERAL_NEVER_INDEXED),
                     None => {
                         out.push(LITERAL_NEVER_INDEXED);
                         self.encode_string(out, name);
@@ -105,11 +105,11 @@ impl Encoder {
 
             match self.table.find(name, header.value()) {
                 Some(index) => {
-                    integer::encode(out, index as u32, 7, INDEXED);
+                    integer::encode(out, index as u64, 7, INDEXED);
                 }
                 None => {
                     match self.table.find_name(name) {
-                        Some(index) => integer::encode(out, index as u32, 6, LITERAL_WITH_INDEXING),
+                        Some(index) => integer::encode(out, index as u64, 6, LITERAL_WITH_INDEXING),
                         None => {
                             out.push(LITERAL_WITH_INDEXING);
                             self.encode_string(out, name);
